requirements.txt
aiomysql==0.2.0
redis==4.2
argparse==1.4.0
betterproto==2.0.0b4
cryptography==41.0.7
dynamic-yaml==1.3.3
fastapi==0.100.0
fastapi-mail==1.4.1
geoip2
grpclib==0.4.5
httpx==0.24.1
ipwhois==1.2.0
mysql-connector-python==8.0.32
py-redis==1.1.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic[email]
pytest==7.4.0
pytest-xdist==2.5.0
pytest-asyncio==0.23.2
python-dateutil==2.8.2
python-pidfile==3.1.1
pytz==2023.3.post1
requests==2.25.1
sqlalchemy==2.0.18
urllib3==1.26.5

